+++ pkg/db/ent/dictionarydetail/dictionarydetail.go
@@ -71,7 +71,7 @@ var (
 	// UpdateDefaultUpdatedAt holds the default value on update for the "updated_at" field.
 	UpdateDefaultUpdatedAt func() time.Time
 	// DefaultStatus holds the default value on creation for the "status" field.
-	DefaultStatus int8
+	DefaultStatus int64
 )
 
 // OrderOption defines the ordering options for the DictionaryDetail queries.
