+++ pkg/db/ent/user/user.go
@@ -92,7 +92,7 @@ var (
 	// UpdateDefaultUpdatedAt holds the default value on update for the "updated_at" field.
 	UpdateDefaultUpdatedAt func() time.Time
 	// DefaultStatus holds the default value on creation for the "status" field.
-	DefaultStatus int8
+	DefaultStatus int64
 	// DefaultSideMode holds the default value on creation for the "side_mode" field.
 	DefaultSideMode string
 	// DefaultBaseColor holds the default value on creation for the "base_color" field.
